Moving base clases to Core namespace
