+++ src/TWidgets/Core/BoxWidget.cs
@@ -3,15 +3,15 @@ using System.Collections.Generic;
 using System.Text;
 using TWidgets.Core.Drawing;
 
-namespace TWidgets.Widgets
+namespace TWidgets.Core
 {
-    public abstract class Box : Widget, IBordeable
+    public abstract class BoxWidget : Widget, IBordeable
     {
         public Border Border { get; private set; }
 
         public int Width { get; set; } = 0;
 
-        public Box(string id) : base(id)
+        public BoxWidget(string id) : base(id)
         {
             this.Border = new Border();
         }
