+++ src/TWidgets/Core/RenderEngine.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using TWidgets.Core.Drawing;
+using TWidgets.Util;
 using TWidgets.Widgets;
 
 namespace TWidgets.Core
