+++ src/TWidgets/Widgets/BulletList.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using TWidgets.Core;
 using TWidgets.Core.Drawing;
 using TWidgets.Util;
