+++ src/TWidgets/Widgets/Message.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
+using TWidgets.Core;
 
 namespace TWidgets.Widgets
 {
