+++ src/TWidgets/Widgets/Separator.cs
@@ -2,10 +2,11 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
+using TWidgets.Core;
 
 namespace TWidgets.Widgets
 {
-    public class Separator : Box
+    public class Separator : BoxWidget
     {
         public Separator(string id) : base(id)
         { }
